+++ src/AuraRouter.php
@@ -6,6 +6,7 @@ namespace Middlewares;
 use Aura\Router\RouterContainer;
 use Middlewares\Utils\Factory;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -21,6 +22,11 @@ class AuraRouter implements MiddlewareInterface
      * @var string Attribute name for handler reference
      */
     private $attribute = 'request-handler';
+
+    /**
+     * @var ResponseFactoryInterface
+     */
+    private $responseFactory;
 
     /**
      * Set the RouterContainer instance.
@@ -40,6 +46,15 @@ class AuraRouter implements MiddlewareInterface
     }
 
     /**
+     * Set the response factory to return the error responses.
+     */
+    public function responseFactory(ResponseFactoryInterface $responseFactory): self
+    {
+        $this->responseFactory = $responseFactory;
+        return $this;
+    }
+
+    /**
      * Process a server request and return a response.
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
@@ -49,17 +64,19 @@ class AuraRouter implements MiddlewareInterface
 
         if (!$route) {
             $failedRoute = $matcher->getFailedRoute();
+            $responseFactory = $this->responseFactory ?: Factory::getResponseFactory();
 
             switch ($failedRoute->failedRule) {
                 case 'Aura\Router\Rule\Allows':
-                    return Factory::createResponse(405)->withHeader('Allow', implode(', ', $failedRoute->allows)); // 405 METHOD NOT ALLOWED
+                    return $responseFactory->createResponse(405)
+                        ->withHeader('Allow', implode(', ', $failedRoute->allows)); // 405 METHOD NOT ALLOWED
                 case 'Aura\Router\Rule\Accepts':
-                    return Factory::createResponse(406); // 406 NOT ACCEPTABLE
+                    return $responseFactory->createResponse(406); // 406 NOT ACCEPTABLE
                 case 'Aura\Router\Rule\Host':
                 case 'Aura\Router\Rule\Path':
-                    return Factory::createResponse(404); // 404 NOT FOUND
+                    return $responseFactory->createResponse(404); // 404 NOT FOUND
                 default:
-                    return Factory::createResponse(500); // 500 INTERNAL SERVER ERROR
+                    return $responseFactory->createResponse(500); // 500 INTERNAL SERVER ERROR
             }
         }