+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2018-08-01
+
+### Added
+
+- New option `responseFactory` to configure a custom factory implementing `Psr\Http\Message\ResponseFactoryInterface`
+
+### Changed
+
+- Upgraded to the final version of PSR-17 `psr/http-factory`
+
 ## [1.0.1] - 2018-02-06
 
 ### Fixed
@@ -102,6 +112,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+
+[1.1.0]: https://github.com/middlewares/aura-router/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/middlewares/aura-router/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/middlewares/aura-router/compare/v0.9.0...v1.0.0
 [0.9.0]: https://github.com/middlewares/aura-router/compare/v0.8.0...v0.9.0